Rename {Side Input -> side_input}
