+++ learning/katas/go/core_transforms/side_input/lesson-info.yaml
@@ -17,25 +17,5 @@
 # under the License.
 #
 
-type: edu
-files:
-- name: test/task_test.go
-  visible: false
-- name: cmd/main.go
-  visible: true
-- name: go.mod
-  visible: false
-- name: go.sum
-  visible: false
-- name: pkg/task/task.go
-  visible: true
-  placeholders:
-  - offset: 1059
-    length: 46
-    placeholder_text: TODO()
-  - offset: 1185
-    length: 83
-    placeholder_text: TODO()
-  - offset: 1273
-    length: 202
-    placeholder_text: TODO()
+content:
+- side_input
