+++ learning/katas/go/core_transforms/side_input/lesson-remote-info.yaml
