+++ learning/katas/go/core_transforms/side_input/side_input/pkg/task/task.go
