+++ learning/katas/go/core_transforms/side_input/side_input/task-info.yaml
@@ -17,5 +17,21 @@
 # under the License.
 #
 
-content:
-- Side Input
+type: edu
+files:
+- name: test/task_test.go
+  visible: false
+- name: cmd/main.go
+  visible: true
+- name: pkg/task/task.go
+  visible: true
+  placeholders:
+  - offset: 1059
+    length: 46
+    placeholder_text: TODO()
+  - offset: 1185
+    length: 83
+    placeholder_text: TODO()
+  - offset: 1273
+    length: 202
+    placeholder_text: TODO()
