+++ learning/katas/go/core_transforms/side_input/side_input/task-remote-info.yaml
