+++ learning/katas/go/core_transforms/side_input/side_input/task.md
